pyproject.toml
[project]
name = "intelligent-tcm-consultation-system"
version = "0.1.0"
description = "MVP for an intelligent TCM consultation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "numpy>=2.2.0",
    "pillow>=11.0.0",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["backend/tests"]
